Update Perryhook.py
Ok so basically when being rate limited it's still using the amount u put so u lose one message for rate limit so we basically just -1 when we get rate limited so u get the exact amount u put instead of losing 1 message cause of the rate limit. (Was trying to find bugs cause I was bored and realized this was a issue).